refactor(callbacks): 使用 Strategy+Registry 模式拆分 _extract_tool_data
将 760 行的 if-elif 分支链分解为 26 个独立提取器函数，通过装饰器注册到
模块级字典。主方法退化为 JSON 解析 + 委托调用，消除单一函数膨胀问题。
